fix: resolve all 30 mypy errors in reachq package (0 errors)
Root causes fixed:
- dijkstra/truncated_dijkstra: return dict[object, float] (not int),
  removing float('inf') into int dict type mismatches
- hopset values: float throughout (matching dijkstra output)
- shortest_path_hopbound: accepts float hopset weights
- closed_form.py: iterate range(n) and typed vertex lists instead of
  g.vertices() (set[object])
- hopset/algorithm: isinstance narrowing for scc_rep indexing
- All 8 previous type: ignore comments in shortest_paths.py removed

reachq/ package: 0 mypy errors (was 30)
scripts/: 40 pre-existing errors in standalone eval scripts (unchanged)

+++ reachq/core/invariants.py
@@ -151,7 +151,7 @@ def assert_partition_correctness(
 
 def assert_distance_approximation(
     graph: WeightedDigraph,
-    hopset: dict[tuple[object, object], int],
+    hopset: dict[tuple[object, object], float],
     source: object,
     epsilon: float,
     max_hops: int,
@@ -209,7 +209,7 @@ def assert_shortcut_set_size_bound(
 
 def assert_hopset_size_bound(
     graph: WeightedDigraph,
-    hopset: dict[tuple[object, object], int],
+    hopset: dict[tuple[object, object], float],
     epsilon: float,
     rho: float,
     msg: str | None = None,
